Split rendered observer crawler tests
